refactor: Extract state logic into useSongs custom hook
Moves all state management logic from SongProvider into a new custom hook, useSongs.
This separates the business logic from the context provider, making the code cleaner, more organized, and the logic potentially reusable.
